# swagger.yaml
openapi: 3.0.3
info:
  title: Reservation Service API
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
...
  /auth/naver/login:
    get:
      tags: [Auth]
      summary: 네이버 로그인 링크 요청
      description: 네이버 OAuth 로그인을 위한 인증 URL을 생성하여 응답한다.
      responses:
        '200':
          description: 로그인 URL 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLoginUrl'
              examples:
                default:
                  value:
                    success: true
                    data:
                      loginUrl: "https://nid.naver.com/oauth2.0/authorize?client_id=..."
                    message: null
        '500':
          description: 서버 에러
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  value:
                    success: false
                    error:
                      code: "INTERNAL_SERVER_ERROR"
                      message: "일시적인 오류가 발생했습니다."
  /auth/naver/callback:
    get:
      tags: [Auth]
      summary: 네이버 로그인 콜백
      description: 네이버에서 전달된 인가 코드를 받아 액세스 토큰 및 사용자 정보를 처리한다.
      parameters:
        - in: query
          name: code
          schema:
            type: string
          required: true
          description: 네이버 OAuth 인가 코드
        - in: query
          name: state
          schema:
            type: string
          required: false
          description: CSRF 방지를 위한 상태 값
      responses:
        '200':
          description: 로그인 처리 성공 (JWT 발급 또는 리다이렉트용 정보)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAuthTokens'
        '400':
          description: 잘못된 요청(코드값 오류 등)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 에러
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponseLoginUrl:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            loginUrl:
              type: string
        message:
          type: string
          nullable: true
    ApiResponseAuthTokens:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/AuthTokens'
        message:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "UNAUTHORIZED"
            message:
              type: string
              example: "인증이 필요합니다."
            details:
              type: object
              nullable: true

    # ---------- 도메인 스키마 ----------
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
...